refactor: move `useUsers()` to entities
