+++ keep-ui/entities/users/model/useUsers.ts
@@ -2,7 +2,7 @@ import { User } from "@/app/(keep)/settings/models";
 import { useHydratedSession as useSession } from "@/shared/lib/hooks/useHydratedSession";
 import { SWRConfiguration } from "swr";
 import useSWRImmutable from "swr/immutable";
-import { useApiUrl } from "./useConfig";
+import { useApiUrl } from "../../../utils/hooks/useConfig";
 import { fetcher } from "utils/fetcher";
 
 export const useUsers = (options: SWRConfiguration = {}) => {
@@ -10,8 +10,8 @@ export const useUsers = (options: SWRConfiguration = {}) => {
   const { data: session } = useSession();
 
   return useSWRImmutable<User[]>(
-    () => (session ? `${apiUrl}/auth/users` : null),
-    (url) => fetcher(url, session?.accessToken),
+    () => (session ? "/auth/users" : null),
+    (url) => fetcher(apiUrl + url, session?.accessToken),
     options
   );
 };
