+++ keep-ui/features/create-or-update-incident/ui/create-or-update-incident-form.tsx
@@ -10,7 +10,7 @@ import {
   SelectItem,
 } from "@tremor/react";
 import { FormEvent, useEffect, useState } from "react";
-import { useUsers } from "utils/hooks/useUsers";
+import { useUsers } from "@/entities/users/model/useUsers";
 import { useIncidentActions } from "@/entities/incidents/model";
 import type { IncidentDto } from "@/entities/incidents/model";
 import { getIncidentName } from "@/entities/incidents/lib/utils";
